`kudune info` doesn't require a running container anymore
